First working version of the regressor class structure.
