+++ sktime_dl/regressors/deeplearning/_cnn.py
@@ -1,4 +1,4 @@
-__author__ = "James Large"
+__author__ = "Withington"
 
 import keras
 import numpy as np
@@ -6,12 +6,11 @@ import pandas as pd
 
 from sklearn.base import RegressorMixin
 
-from sktime.regressors.base import BaseRegressor
 from sktime.utils.validation.supervised import validate_X, validate_X_y
 
+from sktime_dl.regressors.deeplearning._base import BaseDeepRegressor
 
-
-class CNNRegressor(BaseRegressor, RegressorMixin):
+class CNNRegressor(BaseDeepRegressor, RegressorMixin):
     """Time Convolutional Neural Network (CNN).
 
     Adapted from the implementation from Fawaz et. al
@@ -75,29 +74,6 @@ class CNNRegressor(BaseRegressor, RegressorMixin):
         self.avg_pool_size = avg_pool_size
         self.nb_conv_layers = nb_conv_layers
         self.filter_sizes = filter_sizes
-
-    # TODO reuse this function that is in classifiers _base.py
-    def check_and_clean_data(self, X, y=None, input_checks=True):
-        if input_checks:
-            if y is None:
-                validate_X(X)
-            else:
-                validate_X_y(X, y)
-
-        if isinstance(X, pd.DataFrame):
-            if X.shape[1] == 1 and not isinstance(X.iloc[0, 0], pd.Series):
-                raise TypeError(
-                    "Input should either be a 2d numpy array, or a pandas dataframe with a single column of Series objects (networks cannot yet handle multivariate problems")
-            elif X.shape[1] > 1:
-                X = X.to_numpy()
-            else:
-                X = np.asarray([a.values for a in X.iloc[:, 0]])
-
-        if len(X.shape) == 2:
-            # add a dimension to make it multivariate with one dimension
-            X = X.reshape((X.shape[0], X.shape[1], 1))
-
-        return X
 
     def build_model(self, input_shape, nb_classes, **kwargs):
         """
@@ -178,32 +154,3 @@ class CNNRegressor(BaseRegressor, RegressorMixin):
         self.is_fitted_ = True
 
         return self
-
-    def predict(self, X, input_checks=True, **kwargs):
-        """
-        Find regression estimate for all cases in X.
-        Parameters
-        ----------
-        X : array-like or sparse matrix of shape = [n_instances, n_columns]
-            The training input samples.
-            If a Pandas data frame is passed (sktime format)
-            If a Pandas data frame is passed, a check is performed that it only has one column.
-            If not, an exception is thrown, since this classifier does not yet have
-            multivariate capability.
-        input_checks: boolean
-            whether to check the X parameter
-        Returns
-        -------
-        predictions : 1d numpy array
-            array of predictions of each instance
-        """
-        # TODO add check_is_fitted
-
-        print('_cnn_regressor X.shape before:', X.shape) # TODO remove
-        X = self.check_and_clean_data(X, input_checks=input_checks)
-        print('_cnn_regressor X.shape after:', X.shape) # TODO remove
-
-        y_pred = self.model.predict(X, **kwargs)
-        if y_pred.ndim == 1:
-            y_pred.ravel()
-        return y_pred
